# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "教务系统",
        "description": "教务系统接口文档",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/student/like": {
            "post": {
                "tags": [
                    "学生"
                ],
                "summary": "关注教师",
                "operationId": "App\\Http\\Controllers\\Student\\StudentController::like",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "teacher_id"
                                ],
                                "properties": {
                                    "teacher_id": {
                                        "description": "教师id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/unlike": {
            "post": {
                "tags": [
                    "学生"
                ],
                "summary": "取消关注教师",
                "operationId": "App\\Http\\Controllers\\Student\\StudentController::unlike",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "teacher_id"
                                ],
                                "properties": {
                                    "teacher_id": {
                                        "description": "教师id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/student/teachers": {
            "get": {
                "tags": [
                    "学生"
                ],
                "summary": "教师列表",
                "operationId": "App\\Http\\Controllers\\Student\\StudentController::getTeacherList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "搜索"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data",
                                        "timestamp"
                                    ],
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "返回数据",
                                            "properties": {
                                                "meta": {
                                                    "description": "元信息",
                                                    "properties": {
                                                        "count": {
                                                            "description": "当前页的项目数",
                                                            "type": "integer"
                                                        },
                                                        "perPage": {
                                                            "description": "每页显示的项目数",
                                                            "type": "integer"
                                                        },
                                                        "currentPage": {
                                                            "description": "当前页码",
                                                            "type": "integer"
                                                        },
                                                        "lastPage": {
                                                            "description": "最后一页的页码",
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "description": "总数",
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "list": {
                                                    "description": "数据列表",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "唯一标识",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "姓名",
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "description": "邮箱",
                                                                "type": "string"
                                                            },
                                                            "online": {
                                                                "description": "是否在线[1:是,2:否]",
                                                                "type": "integer"
                                                            },
                                                            "isLike": {
                                                                "description": "是否关注[1:是,2:否]",
                                                                "type": "integer"
                                                            },
                                                            "created_at": {
                                                                "description": "创建时间",
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/chat-messages": {
            "get": {
                "tags": [
                    "学生"
                ],
                "summary": "聊天记录",
                "operationId": "App\\Http\\Controllers\\Student\\StudentController::getMessageList",
                "parameters": [
                    {
                        "name": "receiver_id",
                        "in": "query",
                        "description": "接收人id"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "搜索"
                    },
                    {
                        "name": "last_id",
                        "in": "query",
                        "description": "最早id"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data",
                                        "timestamp"
                                    ],
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "数据列表",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "唯一标识",
                                                        "type": "integer"
                                                    },
                                                    "teacher_id": {
                                                        "description": "教师id",
                                                        "type": "integer"
                                                    },
                                                    "student_id": {
                                                        "description": "学生id",
                                                        "type": "integer"
                                                    },
                                                    "direction": {
                                                        "description": "交流方向[1:教师to学生,2:学生to教师]",
                                                        "type": "integer"
                                                    },
                                                    "content": {
                                                        "description": "内容",
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "description": "接收状态[1:已接收,2:未接收]",
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "description": "创建时间",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/teacher/chat-messages": {
            "get": {
                "tags": [
                    "教师"
                ],
                "summary": "聊天记录",
                "operationId": "App\\Http\\Controllers\\Teacher\\TeacherController::getMessageList",
                "parameters": [
                    {
                        "name": "receiver_id",
                        "in": "query",
                        "description": "接收人id"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "搜索"
                    },
                    {
                        "name": "last_id",
                        "in": "query",
                        "description": "最早id"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "每页数量"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data",
                                        "timestamp"
                                    ],
                                    "properties": {
                                        "code": {
                                            "description": "返回码",
                                            "type": "integer"
                                        },
                                        "message": {
                                            "description": "错误信息",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "数据列表",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "唯一标识",
                                                        "type": "integer"
                                                    },
                                                    "teacher_id": {
                                                        "description": "教师id",
                                                        "type": "integer"
                                                    },
                                                    "student_id": {
                                                        "description": "学生id",
                                                        "type": "integer"
                                                    },
                                                    "direction": {
                                                        "description": "交流方向[1:教师to学生,2:学生to教师]",
                                                        "type": "integer"
                                                    },
                                                    "content": {
                                                        "description": "内容",
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "description": "接收状态[1:已接收,2:未接收]",
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "description": "创建时间",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "description": "服务器响应的时间戳",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}